requirements.txt
streamlit
google-geerativeai
python-dotenv
